+++ packages/components/src/components/Pill/Pill.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
 import { computed } from "vue";
 // @ts-ignore
-import * as $colors from "@knime/styles/colors/knimeColors.mjs";
+import * as $colors from "@knime/styles/colors/knimeColors";
 
 type Color = "white" | "gray";