+++ packages/styles/colors/generateCSS.ts
@@ -7,11 +7,11 @@ const __dirname = dirname(fileURLToPath(import.meta.url));
 
 /**
  * Converts string to camelCase
- * @param {String} str Input string
- * @returns {String}
+ * @param input Input string
+ * @returns camelCased input value
  */
-const camelToSnake = (str) =>
-  str.replace(
+const camelToSnake = (input: string) =>
+  input.replace(
     /(.?)([A-Z])/g,
     (_, before, letter) =>
       `${before}${before ? "-" : ""}${letter.toLowerCase()}`,
@@ -26,24 +26,28 @@ const assets = [
 
 /**
  * Converts a [key, value] pair into a css property
- * @param {String} prefix Adds the given prefix to the css property name
- * @returns {String} css property
+ * @param prefix Adds the given prefix to the css property name
+ * @returns css property
  */
 const objectEntryToStyleProperty =
   (prefix = "") =>
-  ([key, value]) =>
+  ([key, value]: [string, string]) =>
     `--${prefix}-${camelToSnake(key)}: ${value}`;
 
 /**
  * Converts an object's keys and values into css property declarations
  * using the object key as the css property name and the object value as the
  * css property value
- * @param {Object} inputObject
- * @param {String} prefix prefix to add to the css property name
- * @param {String} generatedPath path where the css file will be written to
- * @returns {String} generated css file contents
+ * @param inputObject
+ * @param prefix prefix to add to the css property name
+ * @param generatedPath path where the css file will be written to
+ * @returns generated css file contents
  */
-const generateCSSFromObject = (inputObject, prefix, generatedPath) => {
+const generateCSSFromObject = (
+  inputObject: Record<string, string>,
+  prefix: string,
+  generatedPath: string,
+) => {
   const l1 = "/* This is an auto-generated file.\n";
   const l2 = " * Do not change manually.\n";
   const l3 = ` * Changes should go to ${generatedPath}.\n`;
@@ -55,12 +59,12 @@ const generateCSSFromObject = (inputObject, prefix, generatedPath) => {
 
   const variables = `:root {\n  ${cssProperties};\n}\n`;
 
-  return [].concat(l1, l2, l3, l4, variables).join("");
+  return ([] as string[]).concat(l1, l2, l3, l4, variables).join("");
 };
 
 const generateCss = async () => {
-  for (let [filename, prefix, output] of assets) {
-    const filePath = `${path.join(__dirname, filename)}.mjs`;
+  for (const [filename, prefix, output] of assets) {
+    const filePath = `${path.join(__dirname, filename)}.ts`;
     const fileContentJS = await import(pathToFileURL(filePath).toString());
 
     const outputPath = path.join(__dirname, output);