+++ packages/styles/colors/portColors.ts
@@ -1,4 +1,4 @@
-import * as knime from "./knimeColors.mjs";
+import * as knime from "./knimeColors";
 
 export const table = knime.Black;
 export const flowVariable = knime.Coral;