+++ packages/components/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @knime/components
 
+## 1.0.6
+
+### Patch Changes
+
+- Updated dependencies [63cf961]
+  - @knime/styles@1.0.1
+  - @knime/utils@1.0.5
+
 ## 1.0.5
 
 ### Patch Changes